+++ src/main/java/com/vt/valuetogether/domain/team/controller/TeamController.java
@@ -4,7 +4,7 @@ import com.vt.valuetogether.domain.team.dto.reponse.TeamCreateRes;
 import com.vt.valuetogether.domain.team.dto.reponse.TeamDeleteRes;
 import com.vt.valuetogether.domain.team.dto.request.TeamCreateReq;
 import com.vt.valuetogether.domain.team.dto.request.TeamDeleteReq;
-import com.vt.valuetogether.domain.team.service.TeamService;
+import com.vt.valuetogether.domain.team.service.impl.TeamServiceImpl;
 import com.vt.valuetogether.global.response.RestResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -20,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/teams")
 public class TeamController {
 
-    private final TeamService teamService;
+    private final TeamServiceImpl teamService;
 
     @PostMapping
     public RestResponse<TeamCreateRes> createTeam(
